Refactor fields_view_get() to exit if client asked for a non-tree view. | #167477
If the client asks for a non-tree view there's no need to search if there are any configurations so we can exit sooner.

This is a preformance improvement but also seems to fix the related issue. In the issue we found see from the logs that although the client is asking for a 'form' view, we're answering with a tree view. This must happen because there's some bug in the super() call that returns a tree so view_configurator thought it had to intervene. The reason is that view_configurator checked the type value returned by super.

With this change, we first honour the type passed in the request so it will not make view_configurator to act in any way. I think that what was happening is that super() was answering with type=tree but returning a form view in the arch.